refactor: move e2e/recording harness out of prod src into e2e-harness/
None of the control-plane / recording / e2e machinery belongs in the wizard's
production source. Relocate src/lib/ci-driver/ → e2e-harness/ at the repo root
(next to e2e-tests/), and sever every prod coupling:

- Remove the ProgramConfig.e2e field (program-step.ts) and the on-program profile
  (delete posthog-integration/e2e.ts, unwire index.ts). Per-program profiles now
  live in the harness — e2e-harness/profiles.ts, profileFor(programId).
- Add an @e2e-harness/* path alias (tsconfig.build.json + jest moduleNameMapper);
  repoint scripts/tests off @lib/ci-driver.

Result: src/ has ZERO references to the harness, and the published tsdown bundle
contains none of it (previously the ~90-byte profile object shipped). Full suite
(1045 tests, 3 snapshots) passes; real-recording render verified under tsx.
