+++ e2e-harness/action-registry.ts
