+++ src/lib/programs/posthog-integration/index.ts
@@ -20,7 +20,6 @@ import { requestDeepLink } from '@utils/provisioning';
 import { openTrackedLink, withUtm } from '@utils/links';
 import type { CloudRegion } from '@utils/types';
 import { POSTHOG_INTEGRATION_PROGRAM } from './steps.js';
-import { POSTHOG_INTEGRATION_E2E_PROFILE } from './e2e.js';
 import { getContentBlocks } from './content/index.js';
 import { buildCodingAgentPrompt } from './handoff.js';
 
@@ -50,10 +49,6 @@ export const posthogIntegrationConfig: ProgramConfig = {
   id: 'posthog-integration',
   steps: POSTHOG_INTEGRATION_PROGRAM,
   getContentBlocks,
-
-  // E2E test definition — the UI choices `wizard-ci --e2e` makes driving this
-  // flow headlessly. Lives in ./e2e.ts (product knowledge, not harness logic).
-  e2e: POSTHOG_INTEGRATION_E2E_PROFILE,
 
   // Basic integration runs without structured user input; drop wizard_ask
   // so the model can't pop modal prompts mid-run. The runner forwards this
